third_party/WebKit/LayoutTests/imported/wpt/html/webappapis/scripting/events/invalid-uncompiled-raw-handler-compiled-once-expected.txt

CONSOLE ERROR: line 23: Uncaught SyntaxError: Unexpected token }
This is a testharness.js-based test.
FAIL Invalid uncompiled raw handlers should only be compiled once. assert_equals: expected (object) null but got (undefined) undefined
Harness: the test ran to completion.
